Cambiando el metodo /put y experimentando con el metodo /delete

+++ src/routes/product.router.js
@@ -6,14 +6,14 @@ const productManager = new ProductManager();
 
 // Middleware para verificar que los campos necesarios existan
 const validateRequiredFields = (req, res, next) => {
-    const { title, description, code, price, stock, category } = req.body;
-  
-    if (!title || !description || !code || !price || !stock || !category) {
-      return res.status(400).json({ error: "Missing required fields" });
-    }
-  
-    next();
-  };
+  const { title, description, code, price, stock, category } = req.body;
+
+  if (!title || !description || !code || !price || !stock || !category) {
+    return res.status(400).json({ error: "Missing required fields" });
+  }
+
+  next();
+};
 
 router.get("/", async (req, res) => {
   const { limit } = req.query;
@@ -32,30 +32,58 @@ router.get("/:pid", async (req, res) => {
 });
 
 router.post("/", validateRequiredFields, async (req, res) => {
-    const data = req.body;
-    try {
-      const result = await productManager.create(data);
-      res.send(result);
-    } catch (error) {
-      res.status(500).json({ error: "Failed to create product" });
-    }
-  });
+  const data = req.body;
+  try {
+    const result = await productManager.create(data);
+    res.send(result);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to create product" });
+  }
+});
 
 router.put("/:pid", async (req, res) => {
   const { pid } = req.params;
-  const data = req.body;
-  const result = await productManager.update({ ...data, id: pid });
-  res.send(result);
+  const updatedFields = req.body;
+  const existingProduct = await productManager.getById(pid);
+
+  if (!existingProduct) {
+    return res.status(404).send({ message: "Product not found" });
+  }
+
+  const updatedProduct = {
+    ...existingProduct,
+    ...updatedFields,
+    id: existingProduct.id, // Ensure the ID remains unchanged
+  };
+
+  try {
+    const result = await productManager.update(updatedProduct);
+    res.send(result);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to update product" });
+  }
 });
+
 
 router.delete("/:pid", async (req, res) => {
   const { pid } = req.params;
-  const result = await productManager.delete(pid);
-  if (result) {
-    res.send({ message: "Product deleted successfully" });
-  } else {
-    res.status(404).send({ message: "Product not found" });
+  const existingProduct = await productManager.getById(pid);
+
+  if (!existingProduct) {
+    return res.status(404).send({ message: "Product not found" });
+  }
+
+  try {
+    const result = await productManager.delete(pid);
+    if (result) {
+      res.send({ message: "Product deleted successfully" });
+    } else {
+      res.status(500).send({ message: "Failed to delete product 5678" });
+    }
+  } catch (error) {
+    res.status(500).json({ error: "Failed to delete product 1234" });
   }
 });
+
 
 export default router;
